packages/chronicler-lite/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chronicler-lite"
version = "1.1.0"
description = "MemVid-backed storage layer for Chronicler"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "shihwesley"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "chronicler-core",
    "memvid-sdk>=2.0.0",
    "pyyaml",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "watchdog"]

[project.urls]
Homepage = "https://github.com/shihwesley/chronicler"
Repository = "https://github.com/shihwesley/chronicler"

[tool.hatch.build.targets.wheel]
packages = ["src/chronicler_lite"]

[tool.hatch.build]
include = [
    "src/chronicler_lite/**/*.py",
    "src/chronicler_lite/hooks/scripts/*.sh",
]
